Plagiarism_Detector/backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
numpy==1.26.0
scikit-learn==1.3.1
sentence-transformers==2.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
transformers==4.35.0
torch==2.1.0 